refactor(web): remove (home) route group
