+++ apps/web/src/app/(dashboard)/_components/coe-section.tsx
