+++ apps/web/src/app/(dashboard)/_components/key-statistics-section.tsx
