+++ apps/web/src/components/recent-posts.test.tsx
@@ -1,6 +1,6 @@
 import type { SelectPost } from "@sgcarstrends/database";
 import { render, screen } from "@testing-library/react";
-import { RecentPosts } from "@web/app/(dashboard)/(home)/_components/recent-posts";
+import { RecentPosts } from "@web/app/(dashboard)/_components/recent-posts";
 
 vi.mock("@heroui/link", () => ({
   Link: ({ href, children, ...props }: any) => (
